Phase 4
Developed a function to create a circlar matrix to make a smooth focal maximum. Biuld-in function of focalMod is not smooth enough. Terra rasters would be faster, but they do not support matrix for focal analysis for the max function. May need to consider hole filling methods, because some LiDAR files show less canopy density.